Update patch set 3
Patch Set 3:

well, RPM build fails:

https://jenkins.o-ran-sc.org/sandbox/view/All/job/ric-plt-sdl-rpm-autotools-packagecloud-stage-master/3/console

with this error:
08:45:00 configure: WARNING: unrecognized options: --with-rpm-dir
I checked the configure.ac file carefully and I don't see a problem, this has me stumped right now.

But the deb build works in LF, see here:

https://jenkins.o-ran-sc.org/sandbox/view/All/job/ric-plt-sdl-deb-autotools-packagecloud-stage-master/2/console


In all the builds there is an easy opportunity for a minor improvement.  In file debian/control:

    Build-Depends: debhelper (>= 10), pkg-config, libtool, automake, gawk, autoconf, automake, autoconf-archive, make, gcc, libboost-all-dev, libhiredis-dev

Consider replacing "libboost-all-dev" with the three items libboost-filesystem-dev libboost-program-options-dev libboost-system-dev 

That will reduce the amount of time & unnecessary work in the build installing all the packages that "all-dev" brings, wow it's so many and you only need the three I highlighted, please see below

Same goes for README.md which has:

    sudo apt install libboost-all-dev

And please extend the README.md to mention the need for the packages:

   autoconf-archive rpm valgrind

Patch-set: 3
CC: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>